+++ src/s8b.h
@@ -1,17 +1,17 @@
 typedef struct S8bWord {
-    unsigned int selector : 4;
+    unsigned long selector : 4;
     unsigned long values : 60;
 } S8bWord;
 
 typedef struct {
     S8bWord* compressed_indices;
-    unsigned int nz;
-    unsigned int nwords;
+    unsigned long nz;
+    unsigned long nwords;
 } S8bCol;
 
-static int item_width[16] = { 0, 0, 1, 2, 3, 4, 5, 6,
+static long item_width[16] = { 0, 0, 1, 2, 3, 4, 5, 6,
     7, 8, 10, 12, 15, 20, 30, 60 };
-static int group_size[16] = { 240, 120, 60, 30, 20, 15, 12, 10,
+static long group_size[16] = { 240, 120, 60, 30, 20, 15, 12, 10,
     8, 7, 6, 5, 4, 3, 2, 1 };
 static long masks[16] = { 0,
     0,
@@ -30,5 +30,5 @@ static long masks[16] = { 0,
     (1 << 30) - 1,
     ((long)1 << 60) - 1 };
 
-S8bWord to_s8b(int count, int* vals);
-S8bCol col_to_s8b_col(int size, int* col);
+S8bWord to_s8b(long count, long* vals);
+S8bCol col_to_s8b_col(long size, long* col);
